Unify BaseSiteCrawler implementation
